+++ R/bumbl-package.R
@@ -1,0 +1,8 @@
+#' @keywords internal
+#' @aliases bumbl-package
+"_PACKAGE"
+
+## usethis namespace: start
+#' @importFrom lifecycle deprecated
+## usethis namespace: end
+NULL